% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerf.R
\name{predict.rerf}
\alias{predict.rerf}
\title{Prediction of test data using Regression-Enhanced Random Forests.}
\usage{
\method{predict}{rerf}(object, newx)
}
\arguments{
\item{object}{an object of class "rerf", as that created by the function rerf}

\item{newx}{matrix of new values for x at which predictions are to be made
function will abort.}
}
\value{
A vector of predicted values.
}
\description{
Prediction of test data using Regression-Enhanced Random Forests.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
mod <- rerf(x, y, 0.1)
predict(mod, newx = x)
}
