% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerf.R
\name{rerf}
\alias{rerf}
\title{Regression-Enhanced Random Forests}
\usage{
rerf(x, y, lambda, ntree = 500, nodesize = 5, weights = NULL)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{A numeric response vector.}

\item{lambda}{See 'lambda' argument in \code{\link[glmnet:glmnet]{glmnet::glmnet()}}.}

\item{ntree}{See 'ntree' argument in \code{\link[randomForest:randomForest]{randomForest::randomForest()}}.}

\item{nodesize}{See 'nodesize' argument in \code{\link[randomForest:randomForest]{randomForest::randomForest()}}.}

\item{weights}{See 'weights' argument in \code{\link[glmnet:glmnet]{glmnet::glmnet()}}.}
}
\value{
An object with S3 class "rerf"
}
\description{
\code{rerf()} implements Regression-Enhanced Random Forests algorithm (based on
Zhang et al., 2019 paper) for regression.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
mod <- rerf(x, y, 0.1)
}
\references{
Zhang, H., Nettleton, D., & Zhu, Z. (2019). Regression-enhanced
random forests. arXiv preprint
\url{https://doi.org/10.48550/arXiv.1904.10416}.
}
\author{
Umberto Minora \email{umbertofilippo@tiscali.it}, based on the paper
by Zhang et al. (2019).
}
