% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerf.R
\name{rerf}
\alias{rerf}
\title{Regression-Enhanced Random Forests}
\usage{
rerf(x, y, lambda, ...)
}
\arguments{
\item{x}{A numeric matrix of predictors. Requirement: nvars >1;
in other words, x should have 2 or more columns. This is a constraint
of \code{\link[glmnet:glmnet]{glmnet::glmnet()}}.}

\item{y}{A numeric response vector.}

\item{lambda}{See 'lambda' argument in \code{\link[glmnet:glmnet]{glmnet::glmnet()}}.}

\item{...}{other arguments passed to \code{\link[glmnet:glmnet]{glmnet::glmnet()}} and
\code{\link[randomForest:randomForest]{randomForest::randomForest()}}.}
}
\value{
An object with S3 class "rerf"
}
\description{
\code{rerf()} implements Regression-Enhanced Random Forests algorithm (based on
Zhang et al., 2019 paper) for regression.
}
\details{
This function is based on the packages \code{randomForest::randomForest}
and \code{glmnet::glmnet}.
}
\examples{
x <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
mod <- rerf(x, y, lambda = 0.1)
}
\references{
Zhang, H., Nettleton, D., & Zhu, Z. (2019). Regression-enhanced
random forests. arXiv preprint
\url{https://doi.org/10.48550/arXiv.1904.10416}.
}
\author{
Umberto Minora \email{umbertofilippo@tiscali.it}, based on the paper
by Zhang et al. (2019).
}
